+++ packages/eas-cli/src/credentials/ios/api/graphql/queries/AppQuery.ts
@@ -22,7 +22,10 @@ export const AppQuery = {
             }
             ${print(AppFragmentNode)}
           `,
-          { fullName }
+          { fullName },
+          {
+            additionalTypenames: ['App'],
+          }
         )
         .toPromise()
     );